+++ apps/ddaas/src/app/[locale]/(home)/legal/[[...slug]]/page.tsx
@@ -1,14 +1,16 @@
 import { getMDXComponents } from '@/components/mdx-components';
 import { appConfig } from '@/lib/appConfig';
 import { SiteIcon } from '@/lib/site-config';
-import { legalSource } from '@/lib/source';
+import { mdxSourceMap } from '@/lib/source';
 import { NotFoundPage } from '@base-ui/components';
 import { createFumaPage } from '@third-ui/fuma/server';
 
+const sourceKey = 'legal';
 const { Page, generateStaticParams, generateMetadata } = createFumaPage({
-  mdxContentSource: legalSource,
+  sourceKey: sourceKey,
+  mdxContentSource: mdxSourceMap[sourceKey],
   getMDXComponents,
-  mdxSourceDir: appConfig.mdxSourceDir.legal,
+  mdxSourceDir: appConfig.mdxSourceDir[sourceKey],
   githubBaseUrl: appConfig.githubBaseUrl,
   siteIcon: <SiteIcon />,
   FallbackPage: NotFoundPage,