+++ apps/ddaas/src/app/[locale]/docs/[...slug]/page.tsx
@@ -1,15 +1,17 @@
 import { getMDXComponents } from '@/components/mdx-components';
 import { appConfig } from '@/lib/appConfig';
 import { SiteIcon } from '@/lib/site-config';
-import { docsSource } from '@/lib/source';
+import { mdxSourceMap } from '@/lib/source';
 import { NotFoundPage } from '@base-ui/components';
 import { createFumaPage } from '@third-ui/fuma/server';
 import { LLMCopyButton } from '@third-ui/fuma/mdx/toc-base';
 
+const sourceKey = 'docs';
 const { Page, generateStaticParams, generateMetadata } = createFumaPage({
-  mdxContentSource: docsSource,
+  sourceKey: sourceKey,
+  mdxContentSource: mdxSourceMap[sourceKey],
   getMDXComponents,
-  mdxSourceDir: appConfig.mdxSourceDir.docs,
+  mdxSourceDir: appConfig.mdxSourceDir[sourceKey],
   githubBaseUrl: appConfig.githubBaseUrl,
   copyButtonComponent: <LLMCopyButton />,
   siteIcon: <SiteIcon />,